# frontend/src/generated/openapi.yml
openapi: 3.0.3
info:
  title: Merism API
  version: 0.1.0
  description: AI-moderated qualitative research platform.
paths:
...
  /api/custom-reports/:
    get:
      operationId: custom_reports_list
      description: |-
        Model viewset that filters queryset by the caller's current team.

        Usage::

            class StudyViewSet(TeamScopedModelViewSet):
                queryset = Study.objects.all()
                serializer_class = StudySerializer

        Override ``get_queryset()`` to further filter; the default implementation
        already restricts rows to ``team=self.get_team()`` if the model has a
        ``team`` FK.
      parameters:
      - name: limit
        required: false
        in: query
        description: 每页返回的结果数量。
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: 返回结果的起始索引位置。
        schema:
          type: integer
      tags:
      - custom-reports
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCustomReportList'
          description: ''
    post:
      operationId: custom_reports_create
      description: |-
        Model viewset that filters queryset by the caller's current team.

        Usage::

            class StudyViewSet(TeamScopedModelViewSet):
                queryset = Study.objects.all()
                serializer_class = StudySerializer

        Override ``get_queryset()`` to further filter; the default implementation
        already restricts rows to ``team=self.get_team()`` if the model has a
        ``team`` FK.
      tags:
      - custom-reports
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomReport'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CustomReport'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CustomReport'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomReport'
          description: ''
...
  /api/custom-reports/{id}/export_csv/:
    get:
      operationId: custom_reports_export_csv_retrieve
      description: Export report as CSV.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: 标识此 自定义报告 的 UUID 字符串。
        required: true
      tags:
      - custom-reports
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomReport'
          description: ''
...
  /api/custom-reports/{id}/toggle_public/:
    post:
      operationId: custom_reports_toggle_public_create
      description: Toggle public share link.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: 标识此 自定义报告 的 UUID 字符串。
        required: true
      tags:
      - custom-reports
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomReport'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CustomReport'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CustomReport'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomReport'
          description: ''
...
  /api/report-questions/:
    get:
      operationId: report_questions_list
      description: |-
        Model viewset that filters queryset by the caller's current team.

        Usage::

            class StudyViewSet(TeamScopedModelViewSet):
                queryset = Study.objects.all()
                serializer_class = StudySerializer

        Override ``get_queryset()`` to further filter; the default implementation
        already restricts rows to ``team=self.get_team()`` if the model has a
        ``team`` FK.
      parameters:
      - name: limit
        required: false
        in: query
        description: 每页返回的结果数量。
        schema:
          type: integer
      - name: offset
        required: false
        in: query
        description: 返回结果的起始索引位置。
        schema:
          type: integer
      tags:
      - report-questions
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReportQuestionList'
          description: ''
    post:
      operationId: report_questions_create
      description: |-
        Model viewset that filters queryset by the caller's current team.

        Usage::

            class StudyViewSet(TeamScopedModelViewSet):
                queryset = Study.objects.all()
                serializer_class = StudySerializer

        Override ``get_queryset()`` to further filter; the default implementation
        already restricts rows to ``team=self.get_team()`` if the model has a
        ``team`` FK.
      tags:
      - report-questions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportQuestion'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ReportQuestion'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReportQuestion'
        required: true
      security:
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportQuestion'
          description: ''
...
  /api/shared/report/{token}/:
    get:
      operationId: shared_report_retrieve
      description: Public read-only view of a shared report.
      parameters:
      - in: path
        name: token
        schema:
          type: string
        required: true
      tags:
      - shared
      security:
      - cookieAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
  /api/study-insights/{id}/rerun/:
    post:
      operationId: study_insights_rerun_create
      description: Trigger re-generation of insights for this study.
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        description: 标识此 研究洞察 的 UUID 字符串。
        required: true
      tags:
      - study-insights
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudyInsights'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/StudyInsights'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/StudyInsights'
        required: true
      security:
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudyInsights'
          description: ''
...
components:
  schemas:
...
    CustomReport:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        study:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 300
        status:
          allOf:
          - $ref: '#/components/schemas/StatusC42Enum'
          readOnly: true
        ai_synthesis:
          type: string
          readOnly: true
        share_token:
          type: string
          readOnly: true
        is_public:
          type: boolean
        share_url:
          type: string
          readOnly: true
        questions_count:
          type: integer
          readOnly: true
        segments_count:
          type: integer
          readOnly: true
        generated_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        error_message:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - ai_synthesis
      - created_at
      - error_message
      - generated_at
      - id
      - questions_count
      - segments_count
      - share_token
      - share_url
      - status
      - study
      - title
      - updated_at
...
    PaginatedCustomReportList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/CustomReport'
...
    PaginatedReportQuestionList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=400&limit=100
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?offset=200&limit=100
        results:
          type: array
          items:
            $ref: '#/components/schemas/ReportQuestion'
...
    QuestionTypeEnum:
      enum:
      - open_ended
      - multi_select
      - single_select
      - rating
      - ranking
      type: string
      description: |-
        * `open_ended` - Open-ended question
        * `multi_select` - Multi-select question
        * `single_select` - Single-select question
        * `rating` - Rating question
        * `ranking` - Ranking question
...
    ReportQuestion:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        report:
          type: string
          format: uuid
        question_number:
          type: integer
          maximum: 2147483647
          minimum: 0
        title:
          type: string
        question_type:
          $ref: '#/components/schemas/QuestionTypeEnum'
        status:
          allOf:
          - $ref: '#/components/schemas/ReportQuestionStatusEnum'
          readOnly: true
        ai_summary:
          type: string
          readOnly: true
        chart_spec:
          readOnly: true
        themes:
          readOnly: true
        quotes:
          readOnly: true
        segment:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - ai_summary
      - chart_spec
      - created_at
      - id
      - quotes
      - report
      - status
      - themes
      - title
      - updated_at
    ReportQuestionStatusEnum:
      enum:
      - pending
      - generating
      - ready
      - failed
      type: string
      description: |-
        * `pending` - Pending
        * `generating` - Generating
        * `ready` - Ready
        * `failed` - Failed
...
    StatusC42Enum:
      enum:
      - draft
      - generating
      - ready
      - failed
      type: string
      description: |-
        * `draft` - Draft
        * `generating` - Generating
        * `ready` - Ready
        * `failed` - Failed
...
    StudyInsights:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        study:
          type: string
          format: uuid
        status:
          type: string
          readOnly: true
        completed_interviews:
          type: integer
          readOnly: true
        avg_session_minutes:
          type: number
          format: double
          readOnly: true
        interview_topics:
          readOnly: true
        executive_summary:
          type: string
          readOnly: true
        generated_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        error_message:
          type: string
          readOnly: true
        highlights_count:
          type: integer
          readOnly: true
        findings_count:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - avg_session_minutes
      - completed_interviews
      - created_at
      - error_message
      - executive_summary
      - findings_count
      - generated_at
      - highlights_count
      - id
      - interview_topics
      - status
      - study
      - updated_at
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid